+++ permissible/models/__init__.py
@@ -5,7 +5,11 @@ from .role_based.core import (
     build_role_field,
     PermDomainFieldMixin,
 )
-from .role_based.hierarchical import HierarchicalPermDomain
+from .role_based.hierarchical import (
+    HierarchicalPermDomain,
+    walk_ancestor_ids,
+    awalk_ancestor_ids,
+)
 from .role_based.base import BasePermDomain, PermDomainModelMetaclass
 from .metaclasses import AbstractModelMetaclass, ExtraPermModelMetaclass
 from .permissible_mixin import PermissibleMixin